joy pc

+++ resources/views/excel/program_report.blade.php
@@ -236,7 +236,7 @@ function setItem($item,$section_id,$program_setting,$tranche){
     }
 
     $item->qty = $item->jan+$item->feb+$item->mar+$item->apr+$item->may+$item->jun+$item->jul+$item->aug+$item->sep+$item->oct+$item->nov+$item->dece;
-    $item->estimated_budget = (int)$item->qty * str_replace(',', '',(int)$item->unit_cost);
+    $item->estimated_budget = (int)$item->qty * str_replace(',', '',(float)$item->unit_cost);
 
     return $item;
 }
